*: Remove logger
The logger was bad and ugly. We might opt in to a structured logger but
for now its ok. We should focus more on returning good errors.

+++ pkg/libpfs/stat_darwin.go
@@ -7,10 +7,10 @@ import (
 
 // lastAccess gets the atime for darwin devices
 func lastAccess(stat *syscall.Stat_t) time.Time {
-	return time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec))
+	return time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec)
 }
 
 // createTime gets the ctime for darwin devices
 func createTime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(int64(stat.Ctimespec.Sec), int64(stat.Ctimespec.Nsec))
+	return time.Unix(stat.Ctimespec.Sec, stat.Ctimespec.Nsec)
 }
